src/heroku_pushed/requirements.txt
Flask==0.12.2
line-bot-sdk==1.5.0
requests==2.24.0
beautifulsoup4==4.9.3